+++ tests/model_specific/test_llama_cpp.py
