+++ tests/need_credentials/test_azureai_openai.py
@@ -7,7 +7,7 @@ import pytest
 
 from guidance import assistant, gen, models, system, user
 
-from ..models import common_chat_testing
+from ..model_specific import common_chat_testing
 from ..utils import env_or_fail
 
 # Everything in here needs credentials to work
